roopsangamnx_backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
idna==3.7
razorpay==1.4.2
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
<<<<<<< HEAD
# psycopg2-binary==2.9.6
python-dotenv==1.0.0
=======
python-dotenv==1.0.1
psycopg==3.1.19
psycopg-binary==3.1.19
Pillow==10.0.0
whitenoise==6.6.0
setuptools==59.6.0
django-filter==24.2
>>>>>>> release1.1
